initial share handling setup
